PaulterPanTrader/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
yfinance>=0.1.70
websocket-client>=1.2.1
plotly>=5.3.1
dash>=2.0.0
dash-bootstrap-components>=1.0.0
requests>=2.26.0
python-binance>=1.0.15
ta>=0.7.0
scikit-learn>=1.0.0
pyyaml>=6.0
ccxt>=1.60.0
